pyproject.toml
[project]
name = "artefact_calculator_agent"
version = "0.1.0"
description = "Artefact Technical Challenge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai-agents>=0.3.0",
]
